Refactor to avoid circular import
