+++ aodncore/table/__init__.py
@@ -1,0 +1,6 @@
+from .ff import get_field_type, get_tableschema_descriptor
+
+__all__ = [
+    'get_field_type',
+    'get_tableschema_descriptor'
+]
